Add individual GraphicsView
